On new idea, relationship dropdown is defaulted to the current one.
This uses the session variable to store the relationship id when a relationship is loaded (in the relationship controller), and is only cleared when the user goes to home. Although, it may be fine if the variable is never cleared since the new idea can only be called from a relationship show page.